+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to Gogs are documented in this file.
 ### Removed
 
 - The `gogs cert` subcommand. [#8153](https://github.com/gogs/gogs/pull/8153)
+- The `[email] DISABLE_HELO` configuration option. HELO/EHLO is now always sent during SMTP handshake. [#8164](https://github.com/gogs/gogs/pull/8164)
 
 ## 0.14.1